Should be able to finish writing out backup json files to txt file, closing streams.
theres no error handling and no serializing objects, just simply outputting the raw body we get back.

But for the initial goal thats all i need. If i need to get back the score data i can always add in the feature of derserializing it into objects for the purpose of mutations. But for my limited usecase this should be sufficient.